+++ clikt-mordant/src/commonTest/kotlin/com/github/ajalt/clikt/parameters/EnvvarOptionsTest.kt
@@ -1,16 +1,16 @@
 package com.github.ajalt.clikt.parameters
 
-import com.github.ajalt.clikt.core.*
+import com.github.ajalt.clikt.core.BadParameterValue
+import com.github.ajalt.clikt.core.MutuallyExclusiveGroupException
+import com.github.ajalt.clikt.core.context
+import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.parameters.groups.OptionGroup
 import com.github.ajalt.clikt.parameters.groups.cooccurring
 import com.github.ajalt.clikt.parameters.groups.mutuallyExclusiveOptions
 import com.github.ajalt.clikt.parameters.groups.single
 import com.github.ajalt.clikt.parameters.options.*
 import com.github.ajalt.clikt.parameters.types.int
-import com.github.ajalt.clikt.testing.TestCommand
-import com.github.ajalt.clikt.testing.TestSource
-import com.github.ajalt.clikt.testing.defaultLocalization
-import com.github.ajalt.clikt.testing.parse
+import com.github.ajalt.clikt.testing.*
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.data.blocking.forAll
 import io.kotest.data.row
@@ -19,10 +19,6 @@ import kotlin.js.JsName
 import kotlin.test.Test
 
 class EnvvarOptionsTest {
-    private fun <T : CliktCommand> T.withEnv(vararg entries: Pair<String, String?>): T {
-        return context { readEnvvar = { entries.toMap()[it] } }
-    }
-
     @Test
     @JsName("explicit_envvar")
     fun `explicit envvar`() {